backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.26.0
scipy>=1.11.0
scikit-image>=0.22.0
Pillow>=10.1.0
mediapipe>=0.10.0

# Nonlinear Dynamics
nolds>=0.5.2
antropy>=0.1.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
redis>=5.0.0

# Data Processing
pandas>=2.1.0

# Audio Generation (for entrainment)
sounddevice>=0.4.6
soundfile>=0.12.1

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
